+++ CHANGELOG.md
@@ -6,3 +6,11 @@
 
  - *initial alpha*
 
+## [0.1.1]
+
+#### 05/15/24
+
+ - *rule editor updates with state and facts editor*
+ - *minor typing updates that were missed*
+ - *more progressive facts validation*
+